requirements.txt
aiofiles==24.1.0
aiogram==3.27.0
aiogram_dialog==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-solo==2.4.0
frozenlist==1.6.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
loguru==0.7.3
magic-filter==1.0.12
MarkupSafe==3.0.2
maxapi==0.9.18
multidict==6.4.4
packaging==25.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
puremagic==1.30
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
